requirements.txt
edge_tts==7.0.0
fastapi==0.115.11
jinja2==3.1.5
mistralai==1.5.1
pydantic==2.11.0b1
pygame==2.0.1
python-dotenv==1.0.1
Requests==2.32.3
uvicorn==0.34.0
